# openapi.yaml
openapi: 3.1.1
info:
  title: Teebay APIs
  description: Teebay APIs
  version: 1.0.0
servers:
  - url: "{base_url}"
    variables:
      base_url:
        default: http://localhost:5000
        description: Base URL for the API
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BadRequestError:
      type: object
      properties:
        statusCode:
          type: number
          example: 400
        errors:
          type: array
          items:
            type: string
          example: []
        message:
          type: string
          example: Bad Request
    UnauthorizedError:
      type: object
      properties:
        statusCode:
          type: number
          example: 401
        errors:
          type: array
          items:
            type: string
          example: []
        message:
          type: string
          example: Unauthorized
    ForbiddenError:
      type: object
      properties:
        statusCode:
          type: number
          example: 403
        errors:
          type: array
          items:
            type: string
          example: []
        message:
          type: string
          example: Forbidden
    NotFoundError:
      type: object
      properties:
        statusCode:
          type: number
          example: 404
        errors:
          type: array
          items:
            type: string
          example: []
        message:
          type: string
          example: Not Found
...
    InternalServerError:
      type: object
      properties:
        statusCode:
          type: number
          example: 500
        errors:
          type: array
          items:
            type: string
          example: []
        message:
          type: string
          example: Internal Server Error
...
    ERentalPeriod:
      type: string
      enum:
        - DAY
        - WEEK
        - MONTH
    EProductStatus:
      type: string
      enum:
        - AVAILABLE
        - SOLD
        - RENTED
    CreateProductDto:
      type: object
      required:
        - title
        - description
        - categories
        - purchasePrice
        - rentPrice
        - rentalPeriod
      properties:
        title:
          type: string
          example: "iPhone 13 Pro"
        description:
          type: string
          example: "Lightly used iPhone 13 Pro in excellent condition"
        categories:
          type: array
          items:
            type: string
          example: ["Electronics", "Mobile Phones"]
        purchasePrice:
          type: number
          example: 800
        rentPrice:
          type: number
          example: 50
        rentalPeriod:
          $ref: "#/components/schemas/ERentalPeriod"
          example: "MONTH"
    UpdateProductDto:
      type: object
      properties:
        title:
          type: string
          example: "iPhone 13 Pro - Updated"
        description:
          type: string
          example: "Updated description with more details"
        categories:
          type: array
          items:
            type: string
          example: ["Electronics", "Smartphones"]
        purchasePrice:
          type: number
          example: 750
        rentPrice:
          type: number
          example: 45
        rentalPeriod:
          $ref: "#/components/schemas/ERentalPeriod"
          example: "WEEK"
    ProductResponse:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        description:
          type: string
        categories:
          type: array
          items:
            type: string
        purchasePrice:
          type: number
        rentPrice:
          type: number
        rentalPeriod:
          $ref: "#/components/schemas/ERentalPeriod"
        status:
          $ref: "#/components/schemas/EProductStatus"
        viewCount:
          type: number
        owner:
          type: object
          properties:
            id:
              type: number
            userProfile:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProductListItemResponse:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        description:
          type: string
        categories:
          type: array
          items:
            type: string
        purchasePrice:
          type: number
        rentPrice:
          type: number
        rentalPeriod:
          $ref: "#/components/schemas/ERentalPeriod"
        status:
          $ref: "#/components/schemas/EProductStatus"
        viewCount:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProductsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ProductListItemResponse"
        meta:
          type: object
...
paths:
...
  /api/v1/products:
    get:
      summary: Get all products
      tags: [Products]
      parameters:
        - in: query
          name: page
          schema:
            type: number
        - in: query
          name: limit
          schema:
            type: number
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsListResponse"
        "400":
          description: Bad request - invalid query parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        "401":
          description: Unauthorized - not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
    post:
      summary: Create a product
      tags: [Products]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductDto"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "400":
          description: Bad request - validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        "401":
          description: Unauthorized - not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
  /api/v1/products/owner/{ownerId}:
    get:
      summary: Get products by owner
      tags: [Products]
      parameters:
        - in: path
          name: ownerId
          required: true
          schema:
            type: number
        - in: query
          name: page
          schema:
            type: number
        - in: query
          name: limit
          schema:
            type: number
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsListResponse"
        "400":
          description: Bad request - invalid parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        "401":
          description: Unauthorized - not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "404":
          description: Owner not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
  /api/v1/products/{id}/views:
    patch:
      summary: Increment product views
      tags: [Products]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Product views incremented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "400":
          description: Bad request - invalid ID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        "401":
          description: Unauthorized - not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
  /api/v1/products/{id}:
    get:
      summary: Get one product by ID
      tags: [Products]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "400":
          description: Bad request - invalid ID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        "401":
          description: Unauthorized - not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
    patch:
      summary: Update a product
      tags: [Products]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductDto"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "400":
          description: Bad request - validation failed or invalid ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        "401":
          description: Unauthorized - not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "403":
          description: Forbidden - you can only update your own products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
    delete:
      summary: Delete a product
      tags: [Products]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Product deleted
        "400":
          description: Bad request - invalid ID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        "401":
          description: Unauthorized - not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "403":
          description: Forbidden - you can only delete your own products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
...